Move processing functions to utils
To keep app.ts tidy, the functions that were built from the main
processing script are now in Processing.ts in the utils directory.

The import change in fhirdefs/load.ts is to prevent a circular
dependency.
