+++ src/app.ts
@@ -3,23 +3,20 @@
 import path from 'path';
 import fs from 'fs-extra';
 import program from 'commander';
-import cloneDeep from 'lodash/cloneDeep';
-import { importText, FSHTank, RawFSH } from './import';
+import { FSHTank } from './import';
 import { exportFHIR, Package } from './export';
 import { IGExporter } from './ig';
 import { logger, stats, Type } from './utils';
-import { loadDependency, loadCustomResources } from './fhirdefs';
+import { loadCustomResources } from './fhirdefs';
 import { FHIRDefinitions } from './fhirdefs';
 import {
-  filterInlineInstances,
-  filterExampleInstances,
-  filterCapabilitiesInstances,
-  filterVocabularyInstances,
-  filterModelInstances,
-  filterOperationInstances,
-  filterExtensionInstances,
-  filterProfileInstances
-} from './utils';
+  ensureInputDir,
+  ensureOutputDir,
+  readConfig,
+  loadExternalDependencies,
+  fillTank,
+  writeFHIRResources
+} from './utils/Processing';
 import { pad, padStart, sample, padEnd } from 'lodash';
 import chalk from 'chalk';
 
@@ -64,7 +61,12 @@ async function app() {
   const isIgPubContext = path.parse(input).base === 'fsh';
   const outDir = ensureOutputDir(input, program.out, isIgPubContext);
 
-  const config = readConfig(input);
+  let config: any;
+  try {
+    config = readConfig(input);
+  } catch {
+    process.exit(1);
+  }
 
   // Load external dependencies
   const defs = new FHIRDefinitions();
@@ -73,7 +75,13 @@ async function app() {
   // Load custom resources specified in ig-data folder
   loadCustomResources(input, defs);
 
-  const tank = fillTank(input, config);
+  let tank: FSHTank;
+  try {
+    tank = fillTank(input, config);
+  } catch {
+    program.outputHelp();
+    process.exit(1);
+  }
   await Promise.all(dependencyDefs);
 
   // Check for StructureDefinition
@@ -109,155 +117,6 @@ async function app() {
   process.exit(exitCode);
 }
 
-export function ensureInputDir(input: string): string {
-  // If no input folder is specified, set default to current directory
-  if (!input) {
-    input = '.';
-    logger.info('path-to-fsh-defs defaulted to current working directory');
-  }
-
-  // Use fsh/ subdirectory if not already specified and present
-  const fshSubdirectoryPath = path.join(input, 'fsh');
-  if (fs.existsSync(fshSubdirectoryPath)) {
-    input = path.join(input, 'fsh');
-    logger.info('fsh/ subdirectory detected and add to input path');
-  }
-  return input;
-}
-
-export function ensureOutputDir(input: string, output: string, isIgPubContext: boolean): string {
-  if (isIgPubContext) {
-    logger.info(
-      'Current FSH tank conforms to an IG Publisher context. Output will be adjusted accordingly.'
-    );
-  }
-  let outDir = output;
-  if (isIgPubContext && !output) {
-    // When running in an IG Publisher context, default output is the parent folder of the tank
-    outDir = path.join(input, '..');
-    logger.info(`No output path specified. Output to ${outDir}`);
-  } else if (!output) {
-    // Any other time, default output is just to 'build'
-    outDir = path.join('.', 'build');
-    logger.info(`No output path specified. Output to ${outDir}`);
-  }
-  fs.ensureDirSync(outDir);
-  return outDir;
-}
-
-export function readConfig(input: string): any {
-  // Check that package.json exists
-  const packagePath = path.join(input, 'package.json');
-  if (!fs.existsSync(packagePath)) {
-    logger.error('No package.json in FSH definition folder.');
-    process.exit(1);
-  }
-
-  // Parse package.json
-  let config: any;
-  try {
-    config = fs.readJSONSync(packagePath);
-  } catch (e) {
-    logger.error(`The package.json file is not valid JSON: ${packagePath}`);
-    process.exit(1);
-  }
-
-  // Ensure FHIR R4 is added as a dependency
-  const fhirR4Dependency = config.dependencies?.['hl7.fhir.r4.core'];
-  if (fhirR4Dependency !== '4.0.1') {
-    logger.error(
-      'The package.json must specify FHIR R4 as a dependency. Be sure to' +
-        ' add "hl7.fhir.r4.core": "4.0.1" to the dependencies list.'
-    );
-    process.exit(1);
-  }
-  return config;
-}
-
-export function loadExternalDependencies(
-  defs: FHIRDefinitions,
-  config: any
-): Promise<FHIRDefinitions | void>[] {
-  const dependencyDefs: Promise<FHIRDefinitions | void>[] = [];
-  for (const dep of Object.keys(config?.dependencies ?? {})) {
-    dependencyDefs.push(
-      loadDependency(dep, config.dependencies[dep], defs)
-        .then(def => {
-          logger.info(`Loaded package ${dep}#${config.dependencies[dep]}`);
-          return def;
-        })
-        .catch(e => {
-          logger.error(`Failed to load ${dep}#${config.dependencies[dep]}`);
-          logger.error(e.message);
-        })
-    );
-  }
-  return dependencyDefs;
-}
-
-export function fillTank(input: string, config: any): FSHTank {
-  let files: string[];
-  try {
-    files = getFilesRecursive(input);
-  } catch {
-    logger.error('Invalid path to FSH definition folder.');
-    program.outputHelp();
-    process.exit(1);
-  }
-  const rawFSHes = files
-    .filter(file => file.endsWith('.fsh'))
-    .map(file => {
-      const filePath = path.resolve(file);
-      const fileContent = fs.readFileSync(filePath, 'utf8');
-      return new RawFSH(fileContent, filePath);
-    });
-
-  logger.info('Importing FSH text...');
-  const docs = importText(rawFSHes);
-
-  return new FSHTank(docs, config);
-}
-
-export function writeFHIRResources(outDir: string, outPackage: Package, snapshot: boolean) {
-  fs.writeFileSync(
-    path.join(outDir, 'package.json'),
-    JSON.stringify(outPackage.config, null, 2),
-    'utf8'
-  );
-
-  logger.info('Exporting FHIR resources as JSON...');
-  let count = 0;
-  const writeResources = (
-    folder: string,
-    resources: { getFileName: () => string; toJSON: (snapshot: boolean) => any }[]
-  ) => {
-    const exportDir = path.join(outDir, 'input', folder);
-    resources.forEach(resource => {
-      fs.outputJSONSync(path.join(exportDir, resource.getFileName()), resource.toJSON(snapshot), {
-        spaces: 2
-      });
-      count++;
-    });
-  };
-  writeResources('profiles', outPackage.profiles);
-  writeResources('extensions', outPackage.extensions);
-  writeResources('vocabulary', [...outPackage.valueSets, ...outPackage.codeSystems]);
-
-  // Sort instances into appropriate directories
-  const instances = cloneDeep(outPackage.instances); // Filter functions below mutate the argument, so clone what is in the package
-  filterInlineInstances(instances);
-  writeResources('examples', filterExampleInstances(instances));
-  writeResources('capabilities', filterCapabilitiesInstances(instances));
-  writeResources('vocabulary', filterVocabularyInstances(instances));
-  writeResources('models', filterModelInstances(instances));
-  writeResources('operations', filterOperationInstances(instances));
-  writeResources('extensions', filterExtensionInstances(instances));
-  writeResources('profiles', filterProfileInstances(instances));
-  writeResources('resources', instances); // Any instance left cannot be categorized any further so should just be in generic resources
-
-  logger.info(`Exported ${count} FHIR resources as JSON.`);
-}
-
 function getVersion(): string {
   const packageJSONPath = path.join(__dirname, '..', 'package.json');
   if (fs.existsSync(packageJSONPath)) {
@@ -265,15 +124,6 @@ function getVersion(): string {
     return `v${packageJSON.version}`;
   }
   return 'unknown';
-}
-
-function getFilesRecursive(dir: string): string[] {
-  if (fs.statSync(dir).isDirectory()) {
-    const ancestors = fs.readdirSync(dir, 'utf8').map(f => getFilesRecursive(path.join(dir, f)));
-    return [].concat(...ancestors);
-  } else {
-    return [dir];
-  }
 }
 
 function printResults(pkg: Package, isIG: boolean) {
