+++ src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './FSHLogger';
 export * from './MasterFisher';
 export * from './Mixin';
 export * from './InstanceDefinitionUtils';
+export * from './Processing';
